+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased Changes
 * Moved iterator types into `thunderdome::iter`. ([#24])
+* Changed `Index::from_bits` to return `Option<Index>` instead of `Index`, and no longer panic. ([#31])
 
 [#24]: https://github.com/LPGhatguy/thunderdome/issues/24
+[#31]: https://github.com/LPGhatguy/thunderdome/issues/31
 
 ## [0.4.2] - 2021-10-07
 * Fixed miri warning for `Arena::get2_mut`. ([#29])